App 컴포넌트도 파일 분리
